Enable reading and saving of binary responses
Wonderful package! Thanks very much.

When using it recently I had need to check that my webservice was returning a very small binary file (pdf around 10KB). I notice that no matter how often I tried, the VCR would not record the response and so playback was also broken.

A look in the code shows that the http response body was only being json_encoded. This throws an error when there is non-ascii /binary data in the response. By converting the string to a UTF-8 string, json_encode now works and so all the responses get recorded properly.

Hope it helps.